[mod] moves drawers to new drawers folder
