+++ src/components/common/styled-components/ApplicationOptionsList.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { List } from "@mui/material";
 
 import { SettingsListItem } from "../../list-items/SettingsListItem";
-import { HelpDrawer } from "../../help";
+import { HelpDrawer } from "components/drawers";
 import { SendFeedbackListItem } from "../../list-items/SendFeedbackListItem";
 
 export const ApplicationOptionsList = () => {
