+++ src/components/drawers/HelpDrawer/HelpContent/AnnotatorHelpContent.json
