+++ src/components/drawers/index.ts
@@ -1,0 +1,4 @@
+export { AnnotatorDrawer } from "./AnnotatorDrawer";
+export { ToolDrawer } from "./ToolDrawer";
+export { HelpDrawer } from "./HelpDrawer/HelpDrawer";
+export { MainDrawer } from "./MainDrawer";
