--- src/components/help/index.ts
@@ -1,1 +1,0 @@
-export { HelpDrawer } from "./HelpDrawer/HelpDrawer";
